Add retries for cold post-install tests
Some tests require the system to "warm up" before they can meaningfully
be run but they require a first pass of the tests to kickstart the
process. The retry number and interval control this.

This is also useful for flaky tests that fail for reasons we cannot
control e.g. network timeouts, remote servers busy, etc